// src/api/common/types/create.type.ts
export type FakeCreateBankStatementOptions = {
	bsName?: string;
	nameLength?: number;
	folio?: string;
	folioLength?: number;
	processID?: number;
	processIDLength?: number;
	rut?: number;
	rutLength?: number;
	url?: string;
	productID?: string;
	productIDLength?: number;
	period?: string;
	periodLength?: number;
	base64?: boolean;
};